LayoutTests/sputnik/Conformance/08_Types/8.4_The_String_Type/S8.4_A14_T1-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S8.4_A14_T1

PASS Expected parsing failure

TEST COMPLETE
